+++ src/extension.ts
@@ -1,46 +1,37 @@
-import * as vscode from "vscode";
+import * as vscode from 'vscode';
 
-import { CreateBarrelCommandHandler } from "./create-barrel-command-handler";
-import { StartCommandHandler } from "./start-command-handler";
+import { CreateBarrelCommandHandler } from './create-barrel-command-handler';
+import { StartCommandHandler } from './start-command-handler';
 
 let fileSystemWatcher: vscode.FileSystemWatcher | undefined;
 
 export function activate(context: vscode.ExtensionContext) {
   const createBarrelCommand = vscode.commands.registerCommand(
-    "autoBarrel.createBarrel",
+    'autoBarrel.createBarrel',
     CreateBarrelCommandHandler.execute
   );
 
-  const startCommand = vscode.commands.registerCommand(
-    "autoBarrel.start",
-    () => {
-      if (typeof fileSystemWatcher !== "undefined") {
-        vscode.window.showInformationMessage("Auto Barrel is already running.");
-        return;
-      }
-      const watchGlob =
-        vscode.workspace
-          .getConfiguration("autoBarrel")
-          .get<string>("watchGlob") || "**/src/**/*!(.spec).[tj]s";
-      fileSystemWatcher = vscode.workspace.createFileSystemWatcher(
-        watchGlob,
-        false,
-        true,
-        false
-      );
+  const startCommand = vscode.commands.registerCommand('autoBarrel.start', () => {
+    if (typeof fileSystemWatcher !== 'undefined') {
+      vscode.window.showInformationMessage('Auto Barrel is already running.');
+      return;
+    }
+    try {
+      const watchGlob = vscode.workspace.getConfiguration('autoBarrel').get<string>('watchGlob') || '**/src/**/*.[tj]s';
+      fileSystemWatcher = vscode.workspace.createFileSystemWatcher(watchGlob, false, true, false);
       fileSystemWatcher.onDidCreate(StartCommandHandler.handleFileAdded);
       fileSystemWatcher.onDidDelete(StartCommandHandler.handleFileDeleted);
+    } catch (error) {
+      console.error(error);
     }
-  );
+  });
 
-  const stopCommand = vscode.commands.registerCommand("autoBarrel.stop", () => {
-    if (typeof fileSystemWatcher !== "undefined") {
+  const stopCommand = vscode.commands.registerCommand('autoBarrel.stop', () => {
+    if (typeof fileSystemWatcher !== 'undefined') {
       fileSystemWatcher.dispose();
       fileSystemWatcher = undefined;
     } else {
-      vscode.window.showInformationMessage(
-        "Auto Barrel is not running, no action taken."
-      );
+      vscode.window.showInformationMessage('Auto Barrel is not running, no action taken.');
     }
   });
 
@@ -50,7 +41,7 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 export function deactivate() {
-  if (typeof fileSystemWatcher !== "undefined") {
+  if (typeof fileSystemWatcher !== 'undefined') {
     fileSystemWatcher.dispose();
   }
 }